requirements.txt
uvicorn==0.23.2
fastapi==0.109.2
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
cloud-sql-python-connector==1.7.0
pandas==2.0.3
pg8000==1.30.4
scikit-learn==1.3.2
google-cloud-aiplatform==1.42.0
langchain==0.1.7
gcsfs==2024.2.0
nest-asyncio==1.6.0